python/requirements.txt
fastapi
uvicorn
aiohttp
python-dotenv
langchain-core
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-fireworks 